requirements.txt
aiomisc==17.3.41
annotated-types==0.6.0
anyio==4.2.0
asyncio==3.4.3
attrs==23.2.0
azure-core==1.30.0
azure-cosmos==4.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
coverage==7.4.1
Cython==3.0.8
dnspython==2.5.0
docopt==0.6.2
esdbclient==1.0.17
fastapi==0.109.2
fastavro==1.9.4
graphql-core==3.2.3
grpcio==1.60.1
h11==0.14.0
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
logging-journald==0.6.7
memphis-py==1.3.1
mmh3==4.1.0
nats-py==2.7.0
pipreqs==0.4.13
protobuf==4.25.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dotenv==1.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sh==2.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
ulid-py==1.1.0
urllib3==2.2.0
uvicorn==0.27.1
yarg==0.1.9
